+++ public/mix-manifest.json
@@ -1,4 +1,6 @@
 {
-    "/js/app.js": "/js/app.js?id=747c4e65aadfa5a5580f",
-    "/css/app.css": "/css/app.css?id=eadb47cdf1a2f60a198a"
+    "/js/app.js": "/js/app.js?id=c194ec3cdfff2e22d4f1",
+    "/css/app.css": "/css/app.css?id=eadb47cdf1a2f60a198a",
+    "/js/canvas.js": "/js/canvas.js?id=2d3cf71c65d885669366",
+    "/js/html2.js": "/js/html2.js?id=9e3f52e33d30372b099f"
 }
