+++ resources/views/strategy/edit.blade.php
@@ -1,5 +1,8 @@
+
+
 <head>
     <script src="https://html2canvas.hertzen.com/dist/html2canvas.js">    </script>
+    <script src="{{ mix('js/html2.js') }}"></script>
 </head>
 @extends('layouts.app')
 
@@ -13,7 +16,7 @@
     <div class="download__content">
       <div class="canvas_download">
 
-        <div class="download__content__name">
+        <div class="download__content__name" data-name="{{$strategy->name}}">
           <h5>作戦名：
             {{$strategy->name}}
           </h5>
@@ -57,7 +60,7 @@
               
               
       
-              <a href="#" class="download__bttom__on btn-primary" id="download_btn" download>ダウンロード</a>
+              <button href="#" class="download__bttom__on btn-primary" id="download_btn" download>ダウンロード</button>
               <form id='delete_btton' action="{{route('strategy.destroy',['id'=> $strategy->id])}}" method="post">
                 @csrf
                 <button id="del" class="download__bttom__delete btn-danger" data-id="{{$strategy->id}}">削除</button>
