+++ webpack.mix.js
@@ -12,5 +12,7 @@ const mix = require('laravel-mix');
  */
 
 mix.js('resources/js/app.js', 'public/js')
+    .js('resources/js/html2.js','public/js')
+    .js('resources/js/canvas.js','public/js')
     .sass('resources/sass/app.scss', 'public/css')
     .version();
